Changes from PR#7
