--- components/HeaderSingle/index.js
@@ -1,1 +1,0 @@
-export * from './HeaderSingle.js';
